Commit working chapter selection and assignment
Chapter selection now works as expected, searching by book id value we
can get that Book and using getChapterByNumber() find a Chapter by it's
number rather than ID (e.g. users selects Revelation and Chapter 22, 22
is not the id for that Chapter of revelation but only it's number. It's
id is some other number depending on the number of Chapters in the
persistence).